+++ projects/ng-virtual-list/src/lib/enums/snap-to-item-align.ts
@@ -1,0 +1,9 @@
+import { SnapToItemAligns } from "./snap-to-item-aligns";
+
+/**
+ * SnapToItemAlign.
+ * @link https://github.com/DjonnyX/ng-virtual-list/blob/20.x/projects/ng-virtual-list/src/lib/enums/snap-to-item-align.ts
+ * @author Evgenii Alexandrovich Grebennikov
+ * @email djonnyx@gmail.com
+ */
+export type SnapToItemAlign = SnapToItemAligns | 'start' | 'center' | 'end';