+++ projects/ng-virtual-list/src/lib/interfaces/animation-params.ts
@@ -8,6 +8,7 @@
  */
 export interface IAnimationParams {
     scrollToItem: number;
+    snapToItem: number;
     navigateToItem: number;
     navigateByKeyboard: number;
 }