+++ projects/ng-virtual-list/src/lib/interfaces/base-virtual-list-item-component.ts
@@ -4,21 +4,21 @@ import { ISize } from './size';
 import { IRenderVirtualListItem } from '../models/render-item.model';
 
 /**
- * Virtual List Item Interface
+ * IBaseVirtualListItemComponent
  * @link https://github.com/DjonnyX/ng-virtual-list/blob/20.x/projects/ng-virtual-list/src/lib/interfaces/base-virtual-list-item-component.ts
  * @author Evgenii Alexandrovich Grebennikov
  * @email djonnyx@gmail.com
  */
-export abstract class BaseVirtualListItemComponent {
-    abstract get id(): number;
-    abstract regular: boolean;
-    abstract set regularLength(v: string)
-    abstract set item(v: IRenderVirtualListItem | null | undefined);
-    abstract get item(): IRenderVirtualListItem | null | undefined;
-    abstract get itemId(): Id | undefined;
-    abstract set renderer(v: TemplateRef<any> | undefined);
-    abstract get element(): HTMLElement;
-    public abstract getBounds(): ISize;
-    public abstract show(): void;
-    public abstract hide(): void;
+export interface IBaseVirtualListItemComponent {
+    get id(): number;
+    regular: boolean;
+    set regularLength(v: string)
+    set item(v: IRenderVirtualListItem | null | undefined);
+    get item(): IRenderVirtualListItem | null | undefined;
+    get itemId(): Id | undefined;
+    set renderer(v: TemplateRef<any> | undefined);
+    get element(): HTMLElement;
+    getBounds(): ISize;
+    show(): void;
+    hide(): void;
 }